+++ paws.common/NEWS.md
@@ -3,6 +3,7 @@
 * cache `read_ini` for improved performance
 * cache unix os environment variables to improve performance on unix systems.
 * support `web_identity_token_file` in AWS config file thanks to @liuquinlin for implementation.
+* fix `NULL` method `as.POSIXct.default` for older versions of R
 
 # paws.common 0.6.2
 * fix how `read_ini` reads empty profiles from ini files