+++ paws.common/NEWS.md
@@ -1,4 +1,4 @@
-# paws.common 0.6.2.9000
+# paws.common 0.6.3
 * minor performance improvement for `read_ini`
 * cache `read_ini` for improved performance
 * cache unix os environment variables to improve performance on unix systems.